Removed the genVariable' function from GenOps
Now that the C++ generator no longer overrides genVariable', it seems wise to remove it from the GenOps function dictionary.

You can call it via either genVariable or genVariableUnchecked anyway, so genVariable' is just used by the two former functions as implementation.  It can easily be put back into the dictionary later on, if that turns out to be necessary.