Changed tock to use autoconf/automake for its build process
The old Makefile has been removed, and configure.ac and Makefile.am added in its place, along with the four files (AUTHORS, ChangeLog, NEWS and README) that autoconf requires you to have (all currently blank).  tock and tocktest can now be compiled, but the rules for cgtests and compiling the C and C++ files are currently missing.