Changed findAllProcess to be much more efficient
Previously, it was finding the processes in the flow-graph, then using the routes to find the process in question, then calling the examination function on that process.  For large flow trees, the cost of traversing all the routes (to all A.Process!) was too high.

Now it find the processes in the AST (so that it only traverses the AST once) then zips any found processes with the flow-graph (using the route IDs) with Map.intersectionWith, which is much faster.

Fixes #90